Enable task input with Makefile
The Makefile now makes use of a variable "INPUT", which defines the
suffix of our test input. End of test input is marked with EOT (\x04).

The following options are supported
- INPUT is not specified: run and debug will run with user input.
- `INPUT=`: Using large task input from file `day?/input`
- `INPUT=XXX`: Using input from file `day?/input_XXX`

The input files are currently placed next to the src files in
`./src/day?/input???`

Using interaction is not possible, when providing an input.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>